refactor(share): extract useShareFlow hook so UI doesn't import entity action
Move mutation, auto-advance effect, native-share/sheet branching, and all
state-machine logic from ShareButton.tsx into useShareFlow() in
features/share/hooks/. ShareButton is now a thin presentational component.
Add ESLint ignore for useShareFlow.ts to allow the useUserTier deep import
(same rationale as the existing widgets/** exemption — symbol-model barrel
intentionally excludes useUserTier to avoid client bundle pollution).
