+++ src/features/share/index.ts
@@ -10,3 +10,5 @@ export type {
 export { mapAnalysisStatus } from './lib/mapAnalysisStatus';
 export { deriveChartStatus } from './lib/deriveChartStatus';
 export type { DeriveChartStatusInput } from './lib/deriveChartStatus';
+export { useShareFlow } from './hooks/useShareFlow';
+export type { UseShareFlowResult } from './hooks/useShareFlow';
